requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blobfile==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
exceptiongroup==1.1.3
fastapi==0.101.1
filelock==3.12.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
lxml==4.9.3
lz4==4.3.2
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
pycryptodomex==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
PyYAML==6.0.1
pyzstd==0.15.9
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
